Do not allow deletion or modification of documents if the user is not the author (owner)

+++ apps/api/controllers/documents/delete.rb
@@ -32,7 +32,7 @@ module Api::Controllers::Documents
       puts "-- access username:u #{@access.username}"
       puts "-- access valid: #{@access.valid}"
 
-      if @access.valid
+      if @access.valid  && @access.username == params['author_name']
         puts "Access valid"
         delete_document(params)
       else
